Re-org files Move datePicker to module Add Noon/Midnight settings
