+++ index.html
@@ -12,7 +12,7 @@
     <!-- Additional JavaScript & CSS -->
     <link rel="stylesheet" href="css/style.css" />
     <script defer src="./node_modules/@easepick/bundle/dist/index.umd.js"></script>
-    <script defer src="js/slider.js"></script>
+    <script defer src="./js/components/slider.js"></script>
     <script defer src="./node_modules/chart.js/dist/chart.umd.js"></script>
     <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css"
     integrity="sha256-p4NxAoJBhIIN+hmNHrzRCf9tD/miZyoHS5obTRR9BMY="
