+++ js/components/slider.js
