+++ js/models/BirdNet2.4.js
@@ -29,17 +29,17 @@ onmessage = async (e) => {
         DEBUG && console.log("load request to worker");
         const { height, width, location } = JSON.parse(
           fs.readFileSync(
-            path.join(__dirname, `../${version}_model_config.json`),
+            path.join(__dirname, `../../${version}_model_config.json`),
             "utf8"
           )
         );
-        const appPath = "../" + location + "/";
+        const appPath = "../../" + location + "/";
         const list = e.data.list;
         const batch = e.data.batchSize;
         const backend = BACKEND || e.data.backend;
         backend === "webgpu" && require("@tensorflow/tfjs-backend-webgpu");
         let labels;
-        const labelFile = `../labels/V2.4/BirdNET_GLOBAL_6K_V2.4_Labels_en.txt`;
+        const labelFile = `../../labels/V2.4/BirdNET_GLOBAL_6K_V2.4_Labels_en.txt`;
         await fetch(labelFile)
           .then((response) => {
             if (!response.ok) throw new Error("Network response was not ok");
